+++ CHANGELOG.md
@@ -17,11 +17,12 @@
 - (Bugfix) Handle optional taints for Storage Operator
 - (Bugfix) Fix Early Connections for 3.10+
 - (Maintenance) yamlfmt as CI Step
-- (Maintenance) Expose Context in OperatorV2 Item Handler 
+- (Maintenance) Expose Context in OperatorV2 Item Handler
 - (Feature) Improve K8S Mock for UT
 - (Feature) (ML) Introduce basic Conditions
 - (Improvement) Raise memory requests for init containers to 50mi
 - (Feature) (ML) Metadata Service Implementation
+- (Feature) (ML) Storage S3 sidecar implementation
 
 ## [1.2.35](https://github.com/arangodb/kube-arangodb/tree/1.2.35) (2023-11-06)
 - (Maintenance) Update go-driver to v1.6.0, update IsNotFound() checks